+++ src/task-edit.js
@@ -1,8 +1,9 @@
 import {MONTHS} from "./data";
-import {createElement} from "./utils";
+import {Component} from "./component";
 
-export class TaskEdit {
+export class TaskEdit extends Component {
   constructor(task) {
+    super();
     this._title = task.title;
     this._dueDate = task.dueDate;
     this._tags = task.tags;
@@ -11,9 +12,8 @@ export class TaskEdit {
     this._repeatingDays = task.repeatingDays;
     this._isFavorite = task.isFavorite;
     this._isDone = task.isDone;
-
-    this._element = null;
     this._onSubmit = null;
+    this._onSubmitButtonClick = this._onSubmitButtonClick.bind(this);
   }
 
   _onSubmitButtonClick(evt) {
@@ -25,10 +25,6 @@ export class TaskEdit {
 
   set onSubmit(fn) {
     this._onSubmit = fn;
-  }
-
-  get element() {
-    return this._element;
   }
 
   _isRepeating() {
@@ -293,20 +289,6 @@ export class TaskEdit {
     this._element.querySelector(`.card__form`)
         .removeEventListener(`submit`, this._onSubmitButtonClick
           .bind(this));
-  }
-
-  render() {
-    if (this._element) {
-      this._element = null;
-    }
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element = null;
   }
 
 }
